requirements.txt
google-generativeai>=0.3.0
gradio>=4.0.0
python-dotenv>=1.0.0
